chore: fee cleanup (#12941)
- `SponsoredFeePaymentMethod` gets into `aztec.js` , but that's it
(meaning the method can be used from aztec.js, but there's no concept of
where's deployed/how to get the address/bytecode, etc). It's under
`@aztec/aztec.js/fee/testing` to make abundantly clear that this is not
a "production acceptable" approach.
- Similarly to `setupCanonicalL2FeeJuice` there's a utility method that
can be used from the CLI/during sandbox setup to deploy a
`SponsoredFeePaymentContract`. It spits out the address where the
contract is located, just like the test accounts.
- You CAN programatically obtain the address of the "canonical"
`SponsoredFeePaymentContract` via `@aztec/sandbox`, but you CANNOT via
`@aztec/aztec.js`. This is because getting the address implies loading
the contract bytecode and it's not a protocol contract, making imports
messy and tripping the user into making poor decisions in their app
design. If you want to use it in your app, obtain it from the sandbox
output or from an announcement (just like a faucet address, for example)
- This address is only canonical in the sense it's salt is hardcoded to
0 (this lives in `@aztec/constants` under `SPONSORED_FPC_SALT`. For
testnet it should be prefunded! @PhilWindle @alexghr

This PR also builds upon the work done in `aztec.js`, allowing us to
finally get rid of the special handling of account contract deployments
`deploy_account_method.ts` by creating a new `FeePaymentMethod` (that's
not exposed externally!) that allows an account to pay for its own
deployment 😁

---------

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>