chore: Fix for e2e gossip network test
This test was failing with an invalid signature:

```
18:06:18   ● e2e_p2p_network › should rollup txs from all peers
18:06:18
18:06:18     r must be 0 < r < CURVE.n
18:06:18
18:06:18       79 |     const { r, s, v } = signature;
18:06:18       80 |     const recoveryBit = toRecoveryBit(v);
18:06:18     > 81 |     const sig = new secp256k1.Signature(r.toBigInt(), s.toBigInt()).addRecoveryBit(recoveryBit);
18:06:18          |                 ^
18:06:18       82 |     const publicKey = sig.recoverPublicKey(hash.buffer).toHex(false);
18:06:18       83 |     return Buffer.from(publicKey, 'hex');
18:06:18       84 | }
18:06:18
18:06:18       at Signature.assertValidity (../../node_modules/@noble/curves/src/abstract/weierstrass.ts:801:46)
18:06:18       at new Signature (../../node_modules/@noble/curves/src/abstract/weierstrass.ts:782:12)
18:06:18       at recoverPublicKey (../../foundation/dest/crypto/secp256k1-signer/utils.js:81:17)
18:06:18       at recoverAddress (../../foundation/dest/crypto/secp256k1-signer/utils.js:44:23)
18:06:18       at BlockAttestation.getSender (../../stdlib/dest/p2p/block_attestation.js:55:27)
18:06:18           at async Promise.all (index 0)
18:06:18       at Object.<anonymous> (e2e_p2p/gossip_network.test.ts:128:21)
18:06:18
```

The cause was that we were trying to recover the signer from an empty
signature (all zeroes). This is not typically an issue since the p2p
client filters them out (see #12740), but in this test we were reaching
directly to the archiver to retrieve them.

This PR filters out empty signatures before recovering signers for them.

It also includes some extra logging and a test for the validation
service (because why not).

Sample error: http://ci.aztec-labs.com/80ef02336cf535f1