fix: Allow use of local blob sink client
In #12857 an irresponsible developer changed the blob sink client
factory so that the remote `HttpBlobSinkClient` was used if there was a
blob archive API defined or even an L1 chain id, since we have a method
for inferring an archive API just by using the L1 chain id.

This has the unintended effect of using the `HttpBlobSinkClient` pretty
much always, since L1 chain id is pretty much always defined. Even when
running locally against anvil. Which means the archiver no longer works
when running locally, since we end up with no local blob sink.

This commit removes the check for L1 chain id when deciding whether to
use a local client or an http one. We assume that, if working in a
"real" environment, we have an L1 consensus host url, given the archive
API is just meant to be a fallback.