chore: Speed up and deflake sentinel test
The sentinel e2e test was sometimes failing since not enough blocks were
mined in time. We believe this was because the offline validator was
being picked up as proposer, so multiple slots end up being missed.

This PR fixes it by bumping timeout and reducing the number of blocks we
mine. We used to mine so many to make sure that each validator was
picked at least one as proposer, so no matter which one we inspected, we
could assert there was at least one block proposed in its stats. We now
just look for one with proposals and assert on it.