chore: Cron snapshot upload in spartan
Adds a cronjob to the spartan templates that curls the full node and
instructs it to start a snapshot upload to a gcs bucket. The admin api
is exposed via a new service defined on the full node, which should not
get exposed to the outside world based on the firewall rules we have
defined in terraform.