fix: Don't use commander to intercept network flag (#13109)
Whilst using commander to intercept the use of the `--network` seemed
like the correct approach. Doing so meant the `--help` option no longer
worked. This approach simply checks the options and environment without
commander.

We also ensure that Ethereum addresses in the chain configs are not
empty strings as this creashes when parsed.