fix: boolean config helper for cli args works now (#13110)
Using `booleanConfigHelper` without `[value]` in the commander flag was
resulting in all our boolean cli args either using the default (if no
value was provided), or to `false`, because the "value" supplied to the
`parse` function was always `undefined`.

This resuled in, e.g. passing `--p2p-enabled` being interpreted as
false.

Now the expected behavior reigns: `--p2p-enabled` and
`--p2p-enabled=true` and `--p2p-enabled true` all result in the flag
being `true`, but `--p2p-enabled=false` and `--p2p-enabled false` result
in the flag being `false`.

In the `testAccounts` case, since it was defined as `true` by default
for the sandbox, it was picking up that value when parsing for the node,
so no matter what it was true.