fix: Handle proven chain events referring to unseen blocks
The L2BlockStream works by comparing a local and a remote block source.
For simplicity, we had created an L2TipsStore that keeps track of
"local" block tips seen, either in memory or storage.

However, when using the in-memory one, if the block stream reported a
proven block number that the tips store hadn't seen, the tip store would
throw due to not having its block hash when it tried to retrieve it
later.

This fixes it by emitting the block hash along with the number from all
block stream events, and storing them in the tips store.

Fixes #13142