fix!: Validate public setup fns and gas in p2p
Adds the phases and gas validators to the p2p client, so txs are checked
for valid setup public functions and valid gas (wrt the latest block).
This required moving logic from the sequencer to either p2p or stdlib,
and injecting a `ContractDataSource` to the p2p module, in order to
validate setup public functions.

**Breaking change:** `SEQ_ALLOWED_SETUP_FN` env var is now
`TX_PUBLIC_SETUP_ALLOWLIST` (cc @devrel)

Fixes #10958