feat: improved pairing point accumulator (#13226)
Clean up and refactor the use of `aggregation_state` in preparation to
perform pairing point aggregation in all of the places its needed. The
class now has equivalent support for Ultra and Mega.

The main components of this work:
- Use the component-owned methods `set_public` and
`reconstruct_from_public` to define serialization/deserialization
methods for `aggregation_state` rather than relying on one-off
implementations and builder methods
- Avoid explicit use of the underlying witness indices and PairingPoints
(array of group elements) in favor of an `aggregation_state` instance
- Template `aggregation_state` on Builder instead of Curve (since only
bn254 is relevant)

---------

Co-authored-by: ludamad <adam.domurad@gmail.com>