feat: add CapsuleArray::for_each (#13425)
Iterating capsule arrays is becoming more and more common, but it's a
bit clunky - particularly if we also need to delete elements from the
array as we go over it. This introduces `for_each`, which greatly
simplifies this and results in optimal performance.

One of the tests is disabled since it requires a feature from the stdlib
that is not yet present (counting oracle mock calls), but it's only a
performance test so I'm happy to merge as-is.