fix(p2p): persist p2p private key on rollup address change (#13529)
fixes https://github.com/AztecProtocol/aztec-packages/issues/13434

Our onboarding guides don't instruct people to generate a p2p private
key and set it as an environment variable. This is more convenient as
their node generates one and stores it. The issue here is that every
time we change rollup address, the stores get deleted. This means all
nodes generate new p2p keys after each upgrade. This result in lots of
dead ENRs in the distributed hash table.

The new flow for getting the peer id private key is as follows:
 1. Check if we have a peer id private key in the config
 2. If not, check if we have a peer id private key persisted in a file
3. If no file path or data directory is provided or exists, check if we
have a peer id private key in the node's store
4. If not, create a new one, then persist it in a file if a file path or
data directory is provided or in the node's store otherwise

If a private key is written to the node's store, a warning log is
emitted to alert the user to restart their node with a file path or data
directory provided if they want their peer id to persist after a rollup
upgrade.