fix(bb/cmake): LMDB build is NOT the father
When it comes to the lmdb.h folder, the build has nothing to do with it. This explains why this was seemingly pattern following but caused issues. Adding this line only encouraged cmake to clean this up thinking lmdb would build it - when instead it failed, complaining about it missing