fix: Validate txs protocol and vk tree roots (#13891) (#13903)
The `protocolContractTreeRoot` in txs was not checked by nodes when
receiving a tx. This would lead to failures when trying to prove an
epoch that contained such a tx.

This PR adds a check to the tx metadata validator to check the protocol
contract tree root. And since we're at it, we also check the protocol
circuits tree root, so we can fail a tx earlier without having to verify
the entire proof.

This is a cherry-pick of #13891 onto alpha-tesnet