feat(rollup): seed snapshots (#13577)
## Overview

Snapshot the seeds across epochs.

How it currently works:
- If we are in the first epoch, we use a genesis seed = uint256.max
- If the seed for the current epoch has already been calculated, we use
the current seed
- Each time we enter a new epoch, we use the latest seed
- If nothing exists, we use the latest sample seed

In cases where we are looking at the committee for a given block from
the future, we may end up calculated a different committee, as we use
the latest sample seed, which may have changed in the time since.

This PR builds ontop of validator snapshots ( where the validator set
cannot change in size during an epoch ) to make sure that even when we
are querying from the future, and the latest seed has changed, we do not
end up with a different calculated committee.

How it should work:
Seeds are no longer stored for each epoch in EpochData, instead they are
snapshotted.
- If an epoch has not been set up before, it's is either genesis, where
it is given the seed of uint224 max, or it will use the last seen seed
for a given epoch.
- Whenever an epoch is setup, it will calculate if it needs to create a
new seed snapshot for the next epoch to use.
- If the epoch is not setup, it will use the most recent seed snapshot.
( same behaviour, but not we can query into the past and have a
consistent latest snapshot )

fixes: https://github.com/AztecProtocol/aztec-packages/issues/13559