feat: testing contract compilation fails as expected (#13896)
It is desirable to have the ability to test that a given contract fails
with a given compilation error. In this PR I introduce
`noir-projects/noir-contracts-comp-failures` directory in which a
contract package can be defined along with the expected error message
and then in the bootstrap.sh script the contracts are compiled and the
output is checked to contain a given error.

The compilation failure check is being run in CI and I verified that it
fails as expected when the error does not match. Did that in [this
run](https://github.com/AztecProtocol/aztec-packages/actions/runs/14740278893/job/41376379014):
<img width="624" alt="image"
src="https://github.com/user-attachments/assets/a4cbe6be-4421-4bf3-87c4-1f7cacce90c1"
/>

The plan is to test more comp failures in followup PRs.