fix(bb): solve memory blowup, acir::Opcode 7kb => 386 bytes (#14042)
Added a script that can be run after acir generations, until
https://github.com/zefchain/serde-reflection/issues/75 has attention
The issue is that we have a giant enum over these static arrays. In
Rust, they are `Box<... array ...>`, which would be `std::unique_ptr<...
array ...>` in C++. We have an issue open to do just this - but
meanwhile, the more practical move is to use std::shared_ptr to avoid
further changes.

Also uses std::move more to reduce time and allocation.