feat: Add manual rollback command to node admin API (#14106)
Adds a `node_rollbackTo` command to the admin API. This will instruct
the archiver to roll back blocks and messages to the specified L2 block
number, clears the proven block marker, and resets the L1 sync point to
that of the specified L2 block.

In addition, if the target block is behind the last finalized block, we
clear p2p and world state dbs and have them sync from scratch from the
archiver, since these components dont know how to roll back to before a
finalized block.