feat: more profiling (#14142)
- Consolidation of our profiling/timing structs in order to surface more
and more data to the user on "where is time spent" when sending TXs.
- Display of profiling information on both playground and CLI wallet
- General improvements for cli-wallet startup time and usability, trying
to remove the mandatory node or pxe requirements for local-only
commands.
- Removed useless fee estimation default param that forced resimulation
on cli-wallet. Heads up! Fee estimation right now is all but disabled in
both playground and cli-wallet, but at least we're not wasting time on
it. Discussed a bit with @iAmMichaelConnor, and will review soon with
sane defaults