chore: fix update issue + add test to catch some of the offenses (#14284)
Fixes an issue that @spalladino found, as some optimisation works
relying on tstore went a bit too far on writes. The diff on the getter
is minimal, but we also add a bunch of tests here such that we should
have a better change to catch it in the future.

Longer term, we would like to ideally get back to a `view` but, right
now, it is what it is.