feat!: Add client version to ENR (#14352)
Closes #14124

The client version is added to the ENR. The alternative was to add it to
the `Status` message, but as the team suggested, I decided on this
approach.

### Implementation details
The client version is injected into `discV5_service` instead of simply
calling `getPackageVersion()` in the service constructor. The reason
behind this is simple: it seemed more correct to do it this way in case
there is any change in versioning logic in the future. This way, we
ensure that the caller of the P2P service and the P2P service always
have the same version.

### Bugfixes
There was a small bug in `getPackageVersion()`, resulting in the object
being returned (`{"." : "0.8.6"}`} instead of string (version), the PR
also fixes this.

### Refactoring
The ENR generation is extracted into the already existing `generate-enr`
and follows the logic of ENR generation for the bootstrap nodes.

### Bootstrap nodes
I deliberately omitted the client version from bootstrap nodes ENR and
did not change the bootstrap node's ENR with each version change, as
this seems counterproductive and counterintuitive.

The alternative was for Bootstrap nodes to either have a hardcoded
version (e.g., `0` or `0.0.0`) or follow alternative versioning (e.g.,
`boot-v1`). However, this also seemed like a worse idea because handling
boot node versions would require special logic and, if not handled
correctly, could cause confusion when displaying this data.

---------

Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>