feat: return, revert, halt,  and context stack lookups (#13912)
This pr got a little bit oversized in complexity and scope. It
essentially adds the "exit" version of the #13758, adding context-level
constraints for return/revert and error (this the "execution-level"
error which should consolidate any errors propagated from subtraces).

Success will be handled in the next pr...hopefully

### Circuit
* The circuit likely has more columns than it needs, but this can be
re-visited or optimised later.
* New selectors (`sel_enter_call`, `sel_exit_call`, `sel_revert`,
`sel_return`, etc.) and constraints to handle context switching - namely
for nested calls `context.pil`. There are additionally columns to
specify returns from contexts that have a parent (nested context).
* Returndata information is constrained when exiting - (it is likely
incorrect for error but it is TBD whether this matters if the error flag
conditions it's use in the parent context during returndatacopy etc)
* `context_stack.pil` has an actual constraint for its selector.
* Added `instr_length` column in `execution.pil` to increment `next_pc`.
The lookup to the instr_fetching is still pending.

### Simulation
* Added a new `revert` method in the `Execution` class to handle revert
operations, ensuring proper state rollback during errors.
* The `ExecutionEvent` now contains an `error` field, but it's not
populated through execution yet.

### Testing 
* Bunch of new tests (although still not enough). The main one is the
updating of the `constraining` tests that include the lookup check and a
simple CALL -> RETURN context switch

---------

Co-authored-by: esau <152162806+sklppy88@users.noreply.github.com>