feat!: app benches and unified PXE creation (#14504)
This PR promotes the client flows from mere "input creators" to fully
featured benchmarks. This allows us to gather a lot more information on
them, instead of just RAM and proving time. ATM they're run with proving
disabled, so the current script is still used to get stats on bb
performance. ~~The only downside is that we lose information on native
proving stats, but those will be strictly worse than WASM. This change
also means the `build_bench` step on `end-to-end` is now NOT required
for the main bootstrap script, which speeds up fast builds. It is still
kept for the development only `bench_ivc` command in
`barretenberg/cpp`.~~

The `build_bench` step has been moved to right before the actual
benchmarking in the `bench` task starts, so we keep the benefits of CI
cache while not imposing an extra mandatory step for everyone on
`./bootstrap.sh`. We still should consider a "2-tier" benchmarking
approach, or a way for test to upload artifacts to cache, in order to
avoid repeated work.

In order to make this reasonable, the way PXE is initialized has been
unified across server and client (browser) versions, now taking an
`options` object instead of individual flags for logging, store and
other misc configurations.