fix: do not cast numeric to bool (#14593)
For https://github.com/noir-lang/noir/pull/8703

In Noir, casting a number to bool is the same as casting that number to
`u1` first, then checking if that `u1` is 0 or 1. That might be a bit
unexpected, and actually in Rust you can't cast numbers to bool, so it's
going to be disallowed soon.

This PR changes `as bool` to `!= 0` in most cases, though in some
unpacking cases it seems the desired behavior was actually `(value as
u1) != 0`, so there are a couple of those (without these a couple of
tests fail).