feat!: Voting Delegation (#14571)
Fixes #13872.

Does a number of fairly big changes:
- deposits into the rollup is relayed all the way into the governance
- power gained is by default delegated to the rollup deposited into
- allow to update delegation
- let the rollup vote with its power 


We do NOT make any major changes to governance itself, we only update a
little in the underlying libraries, and mainly to have return values and
a function that works on both `storage` and `memory` types.

However, on the rollup and GSE in particular, we are making major
changes.

The fact that we are now depositing into the governance (Which has an
exit delay) means that withdrawing from the GSE is no longer as trivial
as before. We now create a withdrawal from the governance, and must
provide the consumers a way to figure if it was claimed or not, we are
therefore returning additional data in the form of a `withdrawalId` for
that.

At the consumer side (rollup staking) this means that after initiating a
withdrawal there is really 2 timelocks to wait for, the governance and
the rollup specific. We can run those in parallel though, so when both
are done, you can exit.

These changes however mean that the rollup and depositing is much more
linked to governance etc.

Furthermore, when deposits are performed, they will by default increase
the power of the ROLLUP that they were deposited into, e.g., either the
specific instance or the canonical, depending on what was deposited
into. The `withdrawer` can chose a different `delegatee` if they don't
agree.

When it comes to the time of voting, there is a voting power per
address, but as there is a `canonical` there are practically 2 addresses
for the canonical, which it will need to cast votes using to spend all
its power.

To handle time here, and ensure that the rollup is only allowing voting
using the power that was at the time of the vote a lot of accounting is
done using snapshots.

Accounting wise, we have some global account, which is the voting power
for an address etc, but then we also have the accounting for individual
rollups which keeps track of the attester data and how much banace and
who they delegate to. For more on this, look at the `DelegationLib`
which implements this. It don't deal much with the actual attester set,
but all the data linked to them.
Most of the delegation code is very similar to the `votes.sol` from
OpenZeppelin, with the main difference being that this is multiple
accountings within the same contract. The multiple accounting are ONLY
really there because of the automatic moving and the ability for same
attester to be at multiple instances etc.



To allow the rollup itself to vote, we have a function on it that will
cast `yea` votes on a provided `proposalId` if the related proposal was:
- proposed by the governance proposer
- the canonical instance in the eyes of governance proposer is rollup
itself

This have a lot of weird quirks with the "automoving", because you might
get into that position by virtue of automove. However, in that case,
your voting power was previously given to an earlier version (which
likely had the same setup) so you would already have voted same way.
Really, if you want to have any say, delegate to yourself. This is kinda
coming from the fact that we are using the `GovernanceProposer` to make
the proposal, and not the rollup directly, so we cannot just check
`1-1`. This could be fixed, if it is seen as a big issue by having the
rollup itself be the proposer, but that also allows the rollup to just
bypass any of the rules that are in the proposer which are there to
protect gov a bit 🤷.


When going through the accounting, there is often checks that need to be
done against both the specific and the canonical to ensure that we don't
get duplicates, which is kinda pain to look at and reason simply about
balances etc.


# Comment on quality

The quality of the pr is subpar. The naming for the "effective" voting
power etc can be weird, and really it should be cleaned up and have a
good chunk of work added for more testing.

The `User` and `UserLib` should probably also be renamed for sanity. 

Top of head tests to add:
- Better testing that the snapshots work fully as expected
- Exit tests with varying delay durations 
- Exit tests with people calling `finaliseWithdraw` on the governance
ahead of time etc.
- Better checks of where we expect the valances to be, depending on if
there are exited from governance without calls to the rollup etc.

---------

Co-authored-by: Mitchell Tracy <mitchellftracy@gmail.com>