chore: Only validate txs taken from proposals (#14708)
When retrieving transactions in the tx collector, we should only
validate txs that are taken from the proposal.

The rest either come from the node's mempool or request/response, both
of which will already be validated.

There won't be enough time to validate all the txs in a block at the
poit of executing it.