fix: tolerate duplicate proposals (#14732)
We used to only look at payloads that come in through events. 

This was a problem because if the payload was already deployed, we don't
emit an event.

So we need to grab the expected address, and insert it ourselves.

At the same time then, we need to make sure we aren't adding the same
payload to the queue multiple times.

Fix #14694