chore: no row disabling in translator (#14640)
In Sumcheck, the `RowDisablingPolynomial` ensures the full Honk relation
is disabled on the last few rows in the trace where randomness resides.
In the case of Translator, the randomness will reside at different
indices and for simplicity we plan to account for that directly via
relations. As such, this PR provides some general cleanup by reducing
duplication and the ability to not use the RowDisablingPolynomial for
sumcheck.