Update how_to_run_prover.md
After the latest versions, the example `docker-compose.yml` file was incompatible, so I suggest a few changes.
The tag is changed from alpha-testnet to latest because the alpha-testnet tag is no longer used.
The binary run with the `entrypoint` method instead of `command` method, otherwise the containers enter a restart loop because they do not recognize the aztec command. Enviroment variable format is changed acording to this.
Prover and broker data volumes are splitted to avoid conflicts.