chore: Do not mutate config object when starting aztec node
The aztec node createAndSync setup modified the input config object, and
we had code paths (e2e test setup) where we depended on this
modification (`rollupVersion` being required downstream).

This PR changes it so it does not modify the config, and any changes to
the config object are made explicit in the caller.