fix: Limit the number of function names stored in archiver
Limits the number of function selector-name pairs stored in memory in
the archiver, and the maximum length for names. After the limit is hit,
no more values are stored, but the method does not fail.

Also enforces limits on the API method to prevent massive requests to
the node.

Fixes #14257