refactor: use bitmap in outbox (#14988)
Fixes #14964.

When consuming an index in a bucket that have already had a consumption
the gas costs are heavily reduced. This will happen almost every
consumption in practice as they index is used as the key and each bucket
fits 256 indexes.

For the `Outbox.t.sol` test, gas savings looks nice and can be seen
here.

```
╭---------------------------------------------------+-----------------+-------+--------+-------+---------╮
| Function Name                                     | Min             | Avg   | Median | Max   | # Calls |
|---------------------------------------------------+-----------------+-------+--------+-------+---------|
| consume (before)                                  | 28872           | 37526 | 38186  | 60619 | 4888    |
|---------------------------------------------------+-----------------+-------+--------+-------+---------|
| consume (after)                                   | 7874            | 16002 | 16334  | 60667 | 4936    |
╰---------------------------------------------------+-----------------+-------+--------+-------+---------╯
```