feat: kernelless simulation (#15066)
In order to get some advanced simulation features (a collection of
"stuff" we call `simulating simulations`), we have to completely bypass
the kernels since the very nature of what we want to do makes them
invalid to the protocol's eyes. However, we still want to collect side
effects and do as many things as faithfully as possible, so things like
gas estimation don't deviate too much.

This PR is a first step towards that goal:

* Adds a generic utility ported from TXE, that mocks what the kernels do
in order to generate a valid `Tail` input (either for public or rollup).
All sideeffects are currently marked as nonrevertible, but this will be
improved in the future to better reflect reality
* Adds a proxy for the `ContractDataProvider` that allows contract
overrides by both classId and address. This only works for private at
the moment and is a little bit ugly, but improving it requires
refactoring the contract data provider itself and its various levels of
caching
* Adds all the necessary plumbing for this to be invoked from PXE's
`simulateTx`, plus a very basic test that reuses our offchain side
effect emission mechanism to publish some authwit data (still not human
readable). This will be extended so they don't overfit the "message"
(event, note) use case, but any generic output from a contract

---------

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>