feat: reward booster as separate contract (#15230)
Fixes #15213.

The booster is now a separate contract that can be replaced entirely by
updating the reward config.

Note that it practically means that it is possible to brick the rollup
by inserting always reverting booster or by making a booster that simple
consumes all the gas. A separate issue will be made for outlining these,
as the configurability introduced here and in other pr's mean that it is
now a possibility in multiple cases.