chore: construct thread ranges as list of active ranges (#15285)
Before this PR, `construct_ranges_for_equal_content_distribution` only
gives a start and end index for each thread. For example:

```
    std::vector<Range> union_ranges = { { 2, 8 }, { 13, 34 }, { 36, 42 }, { 50, 57 } };

    std::vector<Range> expected_thread_ranges = { { 2, 17 }, { 17, 27 }, { 27, 39 }, { 39, 57 } };
```

And then in `compute_row_evaluations` and `compute_combiner`, for each
thread range, a loops checks whether each index is in the union_ranges.
In the above example, it would call `check_is_active` on indices 2-16
even though we should already know that 8-12 is not active.

This PR makes `construct_ranges_for_equal_content_distribution` return a
list of ranges for each thread, each guaranteed to be active. In the
above example, the expected ranges will be:
```
{ { { 2, 8 }, { 13, 17 } }, { { 17, 27 } }, { { 27, 34 }, { 36, 39 } }, { { 39, 42 }, { 50, 57 } } };
```

Now we can eliminate the `check_is_active` calls because we only iterate
on the active ranges.

See https://github.com/AztecProtocol/barretenberg/issues/1195 for
previous profiling results.