refactor: roundabout temp blocklog values (#15333)
Fixes #15252.

Mixes the blocklog and fee header into archive and temp block log. Then
storing the temp blocklog and preheat the slots.

Savings starting to look sweet. Down to 200K for propose and 400K for
propose with validators.

A good chunk of the diff is a new test to make sure we keep the temps
around for long enough.

---------

Co-authored-by: Mitch <mitchell@aztecprotocol.com>