chore: package vk hash with vk (#15318)
Creates new VKAndHash that contains the VerificationKey and the hash.
The hash is needed, pretty much only in the recursive verifier case,
where we must take the hash and check that its equal to the hash
produced by Fiat-Shamir in Oink. We don't do this check in this PR yet,
so no VKs are changed.