chore!: correctly propagate tail kernel public inputs in the hiding circuit (#15408)
Extract the public inputs directly from the stdlib tail kernel proof in
the hiding circuit and set them public to propagate them to the tube.
(This was previously done by extracting the public inputs from the
native proof which would cause them to be disconnected from those used
in the recursive verifier).

Partially addresses
https://github.com/AztecProtocol/barretenberg/issues/1048