docs: update deployment documentation and remove outdated guides
Revised the "What is a Deployment?" documentation to improve clarity and consistency in the contract immutability table, renaming "Reward Distribution Contract" to "RewardDistributor Contract" and "Proposals Contract" to "GovernanceProposer Contract". Removed outdated guides on "Reacting to upgrades" and "FAQs & Common Issues" to streamline documentation. Updated the system requirements for running a sequencer, changing the CPU recommendation from 8-cores to 4-cores.