chore: Deflake e2e validators sentinel (#15477)
Sample failed run [here](http://ci.aztec-labs.com/bfb52a42988553b2)

The e2e validator sentinel test was sometimes failing because the
history reported for a given validator was too small. This could happen
if some proposals were missed: if a proposal is missed, stats for the
attestors in that slot are not reported, since there was nothing for
them to do, as no block proposal reached them.

This PR fixes it so we wait until _something_ happened for the offline
validator.