chore: stdlib bool internal audit  (#15070)
* Our basic binary ops produce a new gate iff both operands are
witnesses. Most of the implementations contained deeply branched logic
that is now replaced with simple `int` arithmetic that looks more robust
and is much easier to read.

* Added documentation in various places. The constraints are made more
explicit
* Removed redundant assignments
* Re-worked the test suite. Now all basic ops and functions are tested
for all combinations of flags (`is_constant`, `witness_bool`,
`witness_inverted`). Added missing tests, checked various edge cases.
* Removed a couple of tests that seemed less expressive
* Removed `must_imply()` method for a vector of implications, as it is
not used anymore
* Fixed a `normalize()` regression - previously it wouldn't check if
(`witness_inverted == false`) and would normalize a witness `bool_t`
even if the condition is satisfied. It is the reason for the VK changes