feat: transcript can hash objects independently (#15510)
Adds new transcript functionality that allows us to hash the vk (and
soon the accumulator) using an independent buffer in the transcript to
allow for the OriginTags to still be correctly assigned and also having
the hash of the object be independent of the transcript state. The hash
of the object is added to the actual hash buffer that is used to
generate challenges in the protocol.

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>