fix!: aggregate correct nested pairing points in the hiding circuit (#15598)
Prior to this fix, the nested pairing points from the tail kernel were
not being propoerly extracted in the hiding circuit. Instead of
extracting them from the public inputs of the tail decider VK, we were
extracting points from the accumulator VK (whose public inputs are not
meaningful).

Note: this changes the VK only for the hiding circuit which is not
captured by our VK consistency checks