chore!: databus consistency checks in the hiding circuit (#15599)
Uses the new KernelIO mechanism to add databus consistency checks to the
hiding circuit. Shoutout to @kashbrti for figuring out exactly what
needed to be done here.

Note: we probably can actually get away with only performing one
consistency check in the hiding circuit but since its basically free to
do both I'll leave it that way for now. This is consistent with all
other kernels

Closes: https://github.com/AztecProtocol/barretenberg/issues/1397