feat: serde using associated constants + deriving serialization in protocol circuits (#15415)
Updating `Serialize` and `Deserialize` traits to use associated
constants and replacing manual implementations of those traits with
derived ones. Doing these 2 changes in one PR because it saves quite a
lot of time and most of it is very easy to review (most of the changes
are just dropping manual serde implementations, deriving them and
updating some imports).

### On moving stuff around

I had to move the following files to `types::abis`:

- `types::block_header`,
- `types::transaction`,
- `types::state_reference` 

This is because there was this weird interdependency where e.g.
`BlockHeader` needed `Serialize` already derived on some of its struct
members that were in `types::abis` but `BlockHeader` itself couldn't be
placed after `abi` in `types::lib.nr` because some of the code in
`types::abis` needed it to be before as it relied on it having `serde`
already derived. Hence the only solution seemed to be to just move it
all to `abis` as well. If this is considered undesirable I would tackle
that in a followup PR. The `abis` module seems a bit arbitrary now but I
would say it's the kind of tech debt that doesn't matter that much.

### Future plans

In a followup PR I plan on updating `Packable` to use associated
constants.

What also came to my mind is that we might want to derive `Empty`
implementations as well. They all basically look like this:

```
impl Empty for ParityPublicInputs {
    fn empty() -> Self {
        ParityPublicInputs { sha_root: 0, converted_root: 0, vk_tree_root: 0 }
    }
}
```

so one simple solution would be to call `deserialize(...)` on an empty
array.


### Other notes
- This PR replaces
https://github.com/AztecProtocol/aztec-packages/pull/11774
- `serde` tests were AI generated
- Tackled the warning caused by an unnecessary `Eq` trait bound in this
PR by accident as I initially thought my changes introduced it (Nico pls
forgive):
<img width="405" alt="image"
src="https://github.com/user-attachments/assets/ee0d886c-c6b4-4fa7-bd80-bccbceec31ea"
/>

### Compilation slowdown

There is a slowdown in compilation of the token contract but it doesn't
seem to be that significant.

5 runs on mainframe on this PR's branch:

<img width="927" height="346" alt="image"
src="https://github.com/user-attachments/assets/573c34f7-8907-4fbe-9ae5-9df095e3e832"
/>

5 runs on mainframe on `next` branch:

<img width="923" height="339" alt="image"
src="https://github.com/user-attachments/assets/25f281c7-7a46-4ce5-843c-0e27ac59bf3f"
/>

### ~Wen merge~

~Currently somewhat blocked by
https://github.com/noir-lang/noir/issues/9130~