chore: enable kernelless simulation in playground (#15685)
What it says on the tin. Simulation speedup is in the neighborhood of
2x-3x, depending on the particular tx.

While at it discovered a few problematic things that would've prevented
the widespread usage of this feature:

* We lacked the ability of completely overriding message discovery on
contracts. This is very useful on simulated accounts, were the public
key note is completely unused. Avoiding message discovery also speeds up
simulation even more.
* There was a bug when emulating kernel logic. In the event the tx used
the first generated nullifier as nonce generator, it was passed unsiloed
to the public processor.

---------

Co-authored-by: Jan Beneš <janbenes1234@gmail.com>