chore: kernels start with eq and reset (#15734)
As a measure of precaution in case of misconfigured ecc ops in an app
circuit that might attempt, for example, to link ecc ops between apps
and kernels, the subtable of ecc ops in each kernel circuit will start
with an eq and reset op. This cannot happen when the API for creating
ecc ops is used properly as the logic in `goblin_element` ensures an eq
and reset op is present at the end of each `batch_mul`, function also
used for `operator+` and `operator-`. Moreover, the
`add_gates_to_ensure_nonzero_polynomials` makes sure that an eq and
reset exists if no other ecc ops are encountered in the circuit.

---------

Co-authored-by: AztecBot <tech@aztecprotocol.com>
Co-authored-by: federicobarbacovi <171914500+federicobarbacovi@users.noreply.github.com>
Co-authored-by: sergei iakovenko <105737703+iakovenkos@users.noreply.github.com>
Co-authored-by: ludamad <adam.domurad@gmail.com>
Co-authored-by: Suyash Bagad <suyash@aztecprotocol.com>
Co-authored-by: Jonathan Hao <jonathan@aztec-labs.com>
Co-authored-by: ledwards2225 <98505400+ledwards2225@users.noreply.github.com>
Co-authored-by: Raju Krishnamoorthy <krishnamoorthy@gmail.com>
Co-authored-by: notnotraju <raju@aztec-labs.com>
Co-authored-by: Lucas Xia <lucasxia01@gmail.com>
Co-authored-by: Khashayar Barooti <khashayar@aztecprotocol.com>
Co-authored-by: Jean M <132435771+jeanmon@users.noreply.github.com>
Co-authored-by: Alex Gherghisan <alexghr@users.noreply.github.com>
Co-authored-by: Santiago Palladino <spalladino@users.noreply.github.com>
Co-authored-by: Santiago Palladino <santiago@aztec-labs.com>
Co-authored-by: ludamad <domuradical@gmail.com>
Co-authored-by: Sarkoxed <75146596+Sarkoxed@users.noreply.github.com>