feat(avm)!: Poseidon2 memory load and error handling (#15693)
This PR adds the memory aware portion of Poseidon2Permutation so it can
be executed as part of the opcode. This includes error handling for OOB
and invalid tag errors.

This has downstream impacts to alot of the tree op codes whose tests
needed to be updated for the new poseidon2 definition.

---------

Co-authored-by: MirandaWood <miranda@aztecprotocol.com>